requirements.txt
python-dotenv==1.1.0
spotipy==2.25.1
ytmusicapi==1.10.3
